LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/scope-implicit-external-expected.txt


FAIL @scope with external stylesheet assert_equals: expected "1" but got "auto"
